requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==11.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
aiosmtplib==2.0.2  # For email, or use aiosmtplib for async
reportlab==4.0.7  # For PDF generation
openpyxl==3.1.2  # For Excel
pandas==2.1.4
numpy==1.26.0
scikit-learn==1.3.2  # For simple ML categorization
websockets==12.0  # For WebSocket support